+++ src/ClusterManager/joblog_manager.py
@@ -32,7 +32,7 @@ from osUtils import mkdirsAsUser
 from config import config, GetStoragePath
 from DataHandler import DataHandler
 
-from cluster_manager import setup_exporter_thread, manager_iteration_histogram, register_stack_trace_dump, update_file_modification_time
+from cluster_manager import setup_exporter_thread, manager_iteration_histogram, register_stack_trace_dump, update_file_modification_time, record
 
 logger = logging.getLogger(__name__)
 
@@ -47,6 +47,7 @@ def create_log(logdir = '/var/log/dlworkspace'):
 
 
 
+@record
 def extract_job_log(jobId,logPath,userId):
     try:
         dataHandler = DataHandler()
