+++ src/init-scripts/setup_sshd.sh
@@ -24,7 +24,7 @@ function retry {
 }
 
 function setup_sshd {
-    apt-get update && apt-get install -y openssh-server
+    time apt-get update && time apt-get install -y openssh-server
 
     # if "DLWS_HOST_NETWORK" enabled, randomly generate port in range: 40000-49999
     if [ "$DLWS_HOST_NETWORK" = "enable" ];
@@ -37,7 +37,7 @@ function setup_sshd {
     echo "${SSH_PORT}" > ${PROC_DIR}/SSH_PORT
     echo "${POD_IP}" > ${PROC_DIR}/POD_IP
 
-    service ssh restart || exit 1
+    time service ssh restart || exit 1
 }
 
 retry setup_sshd
